Make a separate DB query for each patient MR
Hoping this will help address the `OutOfMemoryError` encountered in beta
testing.  (The previous strategy was to query the DB for *all* patient
MRs for a given report and measure, then filter them in memory to only
those found in the aggregate's subject list.)  Consider batching these
queries if the overhead becomes problematic.