Ensure `supplemental-data` exists
This works around an issue in CQF Ruler where, if the first requests
to `$evaluate-measure` are made simultaneously (e.g., by Link running
with multiple evaluation threads), then all but one of them will fail.
Note that this only affects the very first requests to a fresh
installation of CQF Ruler.