Fix threading issues with in-process measure evaluation (#709)
* Pre-compile with non-null subject

This appears to be required for best performance.  Without it, the
initial batch of evaluations runs somewhat slower (still much better
than with *no* pre-compilation, but an order of magnitude slower than
pre-compiling with a null subject).

* Fix thread-safety issue

`InMemoryFhirRepository` does not appear to be thread-safe.  (Concurrent
calls to its `search` method have the potential to throw
`ConcurrentModificationException`s.)  Therefore, we shouldn't be caching
it across evaluations.  Instead, spin up a new repository and
`R4MeasureService` for each evaluation.  (Not a performance concern; we
can still cache `MeasureEvaluationOptions`, which prevents us from
having to recompile libraries.)

* Explicitly permit `/error` route