Misc improvements to Device generation (#708)
* Removing unnecessary arguments from Validator constructor
Keeping track of ImplementationGuide resources found in NpmPackages within the `Validator` separately from the `PrePopulatedValidationSupport` since `PrePopulatedValidationSupport` doesn't do anything with ImplementationGuide
Loading Device.notes with list of implementation guides that are loaded at the time of validation
Update Device.notes in the report when a report is re-validated
Changing interface for IPatientIdProvider to not return the patient list, since it just loads it into the report context anyways and the return value isn't used.
Change `Device` generated to use `note` for events and concept maps. Much more readable. Conferred with Karen and Dave, who agreed.

* Removing unused constant and unnecessary valueset-timing-abbreviation.json
Not logging parse errors in LenientErrorHandler. We only care if there's an exception, and we're catching and logging those anyways.
Bringing the IG status forward from the original IG.
Parsing resources from packages using a stream instead of loading a string up first.
Handling exception during parsing a little nicer for logging